+++ src/main/kotlin/com/loadingbyte/cinecred/drawer/Common.kt
@@ -11,7 +11,7 @@ import java.awt.Color
 
 val STAGE_GUIDE_COLOR = Color(182, 70, 250)
 val SPINE_GUIDE_COLOR = Color(0, 200, 200)
-val BODY_ELEM_GUIDE_COLOR = Color(130, 50, 0)
+val BODY_CELL_GUIDE_COLOR = Color(130, 50, 0)
 val BODY_WIDTH_GUIDE_COLOR = Color(120, 0, 0)
 val HEAD_TAIL_GUIDE_COLOR = Color(0, 100, 0)